+++ pkg/services/object/server.go
@@ -1,9 +1,11 @@
 package object
 
 import (
+	"bufio"
 	"bytes"
 	"context"
 	"crypto/ecdsa"
+	"crypto/sha256"
 	"encoding/base64"
 	"encoding/binary"
 	"errors"
@@ -17,6 +19,7 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
 	icrypto "github.com/nspcc-dev/neofs-node/internal/crypto"
 	iec "github.com/nspcc-dev/neofs-node/internal/ec"
+	ierrors "github.com/nspcc-dev/neofs-node/internal/errors"
 	igrpc "github.com/nspcc-dev/neofs-node/internal/grpc"
 	inetmap "github.com/nspcc-dev/neofs-node/internal/netmap"
 	iobject "github.com/nspcc-dev/neofs-node/internal/object"
@@ -47,6 +50,7 @@ import (
 	protoencoding "github.com/nspcc-dev/neofs-sdk-go/proto/encoding"
 	protoobject "github.com/nspcc-dev/neofs-sdk-go/proto/object"
 	iprotobuf "github.com/nspcc-dev/neofs-sdk-go/proto/protobuf"
+	"github.com/nspcc-dev/neofs-sdk-go/proto/protobuf/protoscan"
 	"github.com/nspcc-dev/neofs-sdk-go/proto/refs"
 	protosession "github.com/nspcc-dev/neofs-sdk-go/proto/session"
 	protostatus "github.com/nspcc-dev/neofs-sdk-go/proto/status"
@@ -60,6 +64,7 @@ import (
 	grpccodes "google.golang.org/grpc/codes"
 	"google.golang.org/grpc/mem"
 	grpcstatus "google.golang.org/grpc/status"
+	"google.golang.org/protobuf/encoding/protowire"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -145,10 +150,30 @@ type sessions interface {
 type Storage interface {
 	sessions
 
-	// VerifyAndStoreObjectLocally checks whether given object has correct format
-	// and, if so, saves it in the Storage. StoreObject is called only when local
-	// node complies with the container's storage policy.
-	VerifyAndStoreObjectLocally(context.Context, object.Object) error
+	// VerifyObjectHeader checks whether given object header has correct format.
+	//
+	// Requires payload checksum to be of [checksum.SHA256] type.
+	VerifyObjectHeader(context.Context, object.Object) error
+
+	// VerifyObjectPayload makes type-based check of object payload. Called only for
+	// objects of type different from [object.TypeRegular].
+	VerifyObjectPayload(context.Context, object.Object) error
+
+	// StoreObjectLocally saves given in-memory object in the local storage.
+	//
+	// Returns [apistatus.Busy] error if storage is currently overloaded.
+	StoreObjectLocally(ctx context.Context, obj object.Object) error
+
+	// InitLocalObjectWrite initializes object write to the local storage, writes
+	// provided header and returns the stream to write payload. Once payload is
+	// written, stream is closed. On success, object becomes saved.
+	//
+	// Resulting function allows to abort operation in case of problems on caller
+	// side. It is not called multiple times, after [io.Closer.Close] or failed
+	// [io.Writer.Write].
+	//
+	// Returns [ierrors.ErrObjectExists] if object already exists in the storage.
+	InitLocalObjectWrite(ctx context.Context, hdr object.Object, hdrLen uint64, hdrW io.WriterTo) (io.WriteCloser, func(), error)
 
 	// SearchObjects selects up to count container's objects from the given
 	// container matching the specified filters.
@@ -195,6 +220,10 @@ const (
 	maxRespDataChunkSize = maxRespMsgSize * 3 / 4             // 25% to meta, 75% to payload
 	addrMsgSize          = 72                                 // 32 bytes object ID, 32 bytes container ID, 8 bytes protobuf encoding
 	maxObjAddrRespAmount = maxRespDataChunkSize / addrMsgSize // each address is about 72 bytes
+)
+
+const (
+	verifyObjectFailMessage = "failed to verify object"
 )
 
 // Server represents Object Service server that provides object manipulation
@@ -1586,7 +1615,41 @@ func readFirstReplicateV2Request(stream protoobject.ObjectService_ReplicateV2Ser
 	return initPart, nil, nil
 }
 
+var replicateV2RequestScheme = protoscan.MessageScheme{
+	Fields: map[protowire.Number]protoscan.MessageField{
+		protoobject.FieldReplicateV2RequestInit:  protoscan.NewMessageField("init", protoscan.FieldTypeNestedMessage),
+		protoobject.FieldReplicateV2RequestChunk: protoscan.NewMessageField("chunk", protoscan.FieldTypeBytes),
+	},
+}
+
+func getChunkFromReplicateV2Request(req mem.BufferSlice) (iprotobuf.BuffersSlice, *protostatus.Status) {
+	var chunkBuffers iprotobuf.BuffersSlice
+
+	var opts protoscan.ScanMessageOptions
+	opts.InterceptBytes = func(num protowire.Number, buffers iprotobuf.BuffersSlice) error {
+		if num != protoobject.FieldReplicateV2RequestChunk {
+			return protoscan.ErrContinue
+		}
+		chunkBuffers = buffers
+		return nil
+	}
+	opts.InterceptNested = func(protowire.Number, iprotobuf.BuffersSlice) error {
+		return errors.New("non-chunk subsequent message")
+	}
+
+	err := protoscan.ScanMessage(iprotobuf.NewBuffersSlice(req), replicateV2RequestScheme, opts)
+	if err != nil {
+		return iprotobuf.BuffersSlice{}, newBadRequestStatus(err.Error())
+	}
+
+	return chunkBuffers, nil
+}
+
 // ReplicateV2 serves neo.fs.v2.object.ObjectService/ReplicateV2 RPC.
+//
+// ReplicateV2 receives requests of [*protoobject.ReplicateV2Request] or
+// [*mem.BufferSlice] type. ReplicateV2 sends response of
+// [*protoobject.ReplicateV2Response] type.
 func (s *Server) ReplicateV2(stream protoobject.ObjectService_ReplicateV2Server) error {
 	initPart, st, err := readFirstReplicateV2Request(stream)
 	if err != nil {
@@ -1596,19 +1659,7 @@ func (s *Server) ReplicateV2(stream protoobject.ObjectService_ReplicateV2Server)
 		return stream.SendAndClose(&protoobject.ReplicateV2Response{Status: st})
 	}
 
-	recvChunkFn := func() ([]byte, *protostatus.Status, error) {
-		req, err := stream.Recv()
-		if err != nil {
-			return nil, nil, err
-		}
-		chunkPart, ok := req.StreamPart.(*protoobject.ReplicateV2Request_PayloadChunk)
-		if !ok {
-			return nil, newBadRequestStatus("non-chunk subsequent message"), nil
-		}
-		return chunkPart.PayloadChunk, nil, nil
-	}
-
-	objSig, st, err := s.replicate(stream.Context(), initPart.Object, initPart.Signature, initPart.SignObject, recvChunkFn)
+	objSig, st, err := s.replicate(stream.Context(), initPart.Object, initPart.Signature, initPart.SignObject, stream)
 	if err != nil {
 		return err
 	}
@@ -1628,9 +1679,7 @@ func (s *Server) Replicate(ctx context.Context, req *protoobject.ReplicateReques
 		}, nil
 	}
 
-	objSig, st, err := s.replicate(ctx, req.Object, req.Signature, req.SignObject, func() ([]byte, *protostatus.Status, error) {
-		return nil, nil, io.EOF
-	})
+	objSig, st, err := s.replicate(ctx, req.Object, req.Signature, req.SignObject, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -1641,7 +1690,86 @@ func (s *Server) Replicate(ctx context.Context, req *protoobject.ReplicateReques
 	}, nil
 }
 
-func (s *Server) replicate(ctx context.Context, objMsg *protoobject.Object, sig *refs.Signature, signObject bool, recvChunkFn func() ([]byte, *protostatus.Status, error)) ([]byte, *protostatus.Status, error) {
+func newWrongReplicatedObjectPayloadLengthStatus() *protostatus.Status {
+	return newInternalServerErrorStatus(fmt.Sprintf("%s: %v", verifyObjectFailMessage, putsvc.ErrWrongPayloadSize))
+}
+
+func newStoreLocalObjectFailureStatus(cause error) *protostatus.Status {
+	return newInternalServerErrorStatus(fmt.Sprintf("failed to store object locally: %v", cause))
+}
+
+var bufIOWriterPool = sync.Pool{
+	New: func() any {
+		return bufio.NewWriterSize(nil, 256<<10)
+	},
+}
+
+// Wraps and returns w error along with errWriteStream.
+func readReplicatedObjectPayload(payloadLen uint64, stream grpc.ServerStream, w io.Writer) ([]byte, uint64, *protostatus.Status, error) {
+	bw := bufIOWriterPool.Get().(*bufio.Writer)
+	bw.Reset(w)
+	defer func() {
+		bw.Reset(nil) // deref w
+		bufIOWriterPool.Put(bw)
+	}()
+
+	var totalLen uint64
+	h := sha256.New()
+
+	for {
+		var reqBuffers mem.BufferSlice
+		err := stream.RecvMsg(&reqBuffers)
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				err = bw.Flush()
+				if err != nil {
+					return nil, 0, nil, fmt.Errorf("%w: %w", errWriteStream, err)
+				}
+				return h.Sum(nil), totalLen, nil, nil
+			}
+			return nil, 0, nil, err
+		}
+
+		chunkBuffers, st := getChunkFromReplicateV2Request(reqBuffers)
+		if st != nil {
+			reqBuffers.Free()
+			return nil, 0, st, nil
+		}
+
+		chunkLen := chunkBuffers.Len()
+		if chunkLen == 0 {
+			reqBuffers.Free()
+			return nil, 0, newBadRequestStatus("empty payload chunk"), nil
+		}
+
+		if totalLen+uint64(chunkLen) > payloadLen {
+			reqBuffers.Free()
+			return nil, 0, newWrongReplicatedObjectPayloadLengthStatus(), nil
+		}
+
+		// Chunk length is limited by max data frame used in gRPC. By default, it is
+		// 16KB. Writing a lot of such chunks can lead to a large number of syscalls.
+		// Bigger frames (https://github.com/nspcc-dev/neofs-node/issues/4155) could be
+		// used as a solution.
+		_, err = chunkBuffers.WriteTo(bw)
+		if err != nil {
+			reqBuffers.Free()
+			return nil, 0, nil, fmt.Errorf("%w: %w", errWriteStream, err)
+		}
+
+		_, err = chunkBuffers.WriteTo(h)
+
+		reqBuffers.Free()
+
+		if err != nil { // not expected to ever happen
+			return nil, 0, newInternalServerErrorStatus(fmt.Sprintf("write to hash failure: %v", err)), nil
+		}
+
+		totalLen += uint64(chunkLen)
+	}
+}
+
+func (s *Server) replicate(ctx context.Context, objMsg *protoobject.Object, sig *refs.Signature, signObject bool, stream grpc.ServerStream) ([]byte, *protostatus.Status, error) {
 	if objMsg.ObjectId == nil || len(objMsg.ObjectId.Value) == 0 {
 		return nil, newBadRequestStatus("ID field is missing/empty in the object field"), nil
 	}
@@ -1739,52 +1867,120 @@ func (s *Server) replicate(ctx context.Context, objMsg *protoobject.Object, sig 
 		return nil, newAccessDeniedStatus("client does not match the object's storage policy"), nil
 	}
 
-	// TODO(@cthulhu-rider): avoid decoding the object completely
 	obj, err := objectFromMessage(objMsg)
 	if err != nil {
 		return nil, newBadRequestStatus(fmt.Sprintf("invalid object field: %v", err)), nil
 	}
 
-	// TODO: avoid full buffering, copy  to local storage stream directly instead
-	payload := slices.Grow(objMsg.Payload, int(hdr.PayloadLength)-len(objMsg.Payload))
-	for {
-		chunk, st, err := recvChunkFn()
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return nil, nil, err
-		}
-		if st != nil {
-			return nil, st, nil
-		}
-
-		if len(chunk) == 0 {
-			return nil, newBadRequestStatus("empty payload chunk"), nil
-		}
-
-		payload = append(payload, chunk...)
-	}
-	obj.SetPayload(payload)
-
-	err = s.storage.VerifyAndStoreObjectLocally(ctx, *obj)
+	err = s.storage.VerifyObjectHeader(ctx, *obj)
 	if err != nil {
-		if errors.Is(err, apistatus.ErrBusy) {
-			return nil, apistatus.FromError(err), nil
-		}
-		return nil, newInternalServerErrorStatus(fmt.Sprintf("failed to verify and store object locally: %v", err)), nil
+		return nil, newInternalServerErrorStatus(fmt.Sprintf("%s: %v", verifyObjectFailMessage, err)), nil
 	}
 
-	var objSig []byte
-	if signObject {
-		objSig, err = s.metaInfoSignature(*obj)
+	makeOKResult := func() ([]byte, *protostatus.Status, error) {
+		if !signObject {
+			return nil, nil, nil
+		}
+
+		objSig, err := s.metaInfoSignature(*obj)
 		if err != nil {
 			return nil, newInternalServerErrorStatus(fmt.Sprintf("failed to sign object meta information: %v", err)), nil
 		}
+
+		// nil status corresponds to OK
+		return objSig, nil, nil
 	}
 
-	// nil status corresponds to OK
-	return objSig, nil, nil
+	var gotPayloadLen uint64
+	var gotHash []byte
+
+	isStorageStream := stream != nil && obj.Type() == object.TypeRegular
+	var storageStreamAbortFn func()
+	var storageStreamCloser io.Closer
+
+	if isStorageStream {
+		// TODO: can be optimized from two sides:
+		//  1. header structure decoding can be done without unmarshaling (e.g. via protoscan funcs)
+		//  2. since header is already serialized in the original request, io.WriterTo can be implemented around it
+		var storageStream io.WriteCloser
+		storageStream, storageStreamAbortFn, err = s.storage.InitLocalObjectWrite(ctx, *obj, uint64(obj.HeaderLen()), iobject.WriterTo(*obj))
+		if err != nil {
+			if errors.Is(err, ierrors.ErrObjectExists) {
+				return makeOKResult()
+			}
+			return nil, newStoreLocalObjectFailureStatus(fmt.Errorf("init stream: %w", err)), nil
+		}
+
+		var st *protostatus.Status
+		gotHash, gotPayloadLen, st, err = readReplicatedObjectPayload(obj.PayloadSize(), stream, storageStream)
+		if err != nil || st != nil {
+			if errors.Is(err, errWriteStream) {
+				return nil, newStoreLocalObjectFailureStatus(err), nil
+			}
+			storageStreamAbortFn()
+			return nil, st, err
+		}
+
+		storageStreamCloser = storageStream
+	} else if stream != nil {
+		if hdr.PayloadLength < uint64(len(objMsg.Payload)) { // also prevents Grow() panic below
+			return nil, newWrongReplicatedObjectPayloadLengthStatus(), nil
+		}
+
+		buf := bytes.NewBuffer(objMsg.Payload)
+		buf.Grow(int(hdr.PayloadLength) - len(objMsg.Payload))
+
+		var st *protostatus.Status
+		gotHash, gotPayloadLen, st, err = readReplicatedObjectPayload(obj.PayloadSize(), stream, buf)
+		if err != nil || st != nil {
+			return nil, st, err
+		}
+
+		obj.SetPayload(buf.Bytes())
+	} else { // Replicate(V1)
+		gotPayloadLen = uint64(len(obj.Payload()))
+		h := sha256.Sum256(obj.Payload())
+		gotHash = h[:]
+	}
+
+	if obj.PayloadSize() != gotPayloadLen {
+		if isStorageStream {
+			storageStreamAbortFn()
+		}
+		return nil, newWrongReplicatedObjectPayloadLengthStatus(), nil
+	}
+
+	if obj.Type() != object.TypeRegular {
+		err = s.storage.VerifyObjectPayload(ctx, *obj)
+		if err != nil {
+			return nil, newInternalServerErrorStatus(fmt.Sprintf("%s: %v", verifyObjectFailMessage, err)), nil
+		}
+	}
+
+	// checksum must be only SHA256, this was checked above
+	if !bytes.Equal(gotHash[:], hdr.GetPayloadHash().GetSum()) {
+		if isStorageStream {
+			storageStreamAbortFn()
+		}
+		return nil, newInternalServerErrorStatus(verifyObjectFailMessage + ": payload SHA-256 checksum mismatch"), nil
+	}
+
+	if isStorageStream {
+		err = storageStreamCloser.Close()
+		if err != nil {
+			return nil, newStoreLocalObjectFailureStatus(fmt.Errorf("close stream: %w", err)), nil
+		}
+	} else {
+		err = s.storage.StoreObjectLocally(ctx, *obj)
+		if err != nil {
+			if errors.Is(err, apistatus.ErrBusy) {
+				return nil, apistatus.FromError(err), nil
+			}
+			return nil, newStoreLocalObjectFailureStatus(err), nil
+		}
+	}
+
+	return makeOKResult()
 }
 
 func (s *Server) signSearchResponse(body *protoobject.SearchV2Response_Body, err error, req *protoobject.SearchV2Request) *protoobject.SearchV2Response {